Add app service to run the app
This means that the developer doesn’t have to run npm start themselves. Docker compose will start the database service, use the populate-database service to import the data, and start the application.

Changed the hostname for the database in the application connect string as it needs to connect via the Docker hostname. This will stop it connecting outside Docker.