Add pgadmin4 service
Exposes pgadmin4 on port 5050. This is running in desktop mode, so that pgadmin4 login credentials are not required. I believe this is okay, because it will only be available from the local host. It should not be run in desktop mode when it is deployed to a web server.

See [Container Deployment docs](https://www.pgadmin.org/docs/pgadmin4/development/container_deployment.html)

pgadmin4 is configured with the server connection, but I’ve not been successful supplying the password, so this must be entered manually. See `database.env` for the password.